[None][chore] Move MLA module to separate file
