+++ tensorrt_llm/_torch/custom_ops/__init__.py
@@ -11,9 +11,9 @@ from .torch_custom_ops import BufferKind, bmm_out
 from .trtllm_gen_custom_ops import fp8_block_scale_moe_runner
 from .userbuffers_custom_ops import add_to_ub, copy_to_userbuffers, matmul_to_ub
 
-# Attention custom ops (attn_custom_op_inplace, mla_custom_op_inplace) are defined in
-# modules.attention and must be imported from there. They are not re-exported here to
-# avoid circular imports: custom_ops must not depend on modules.attention.
+# Attention custom ops are defined in modules.attention, and MLA custom ops are
+# defined in modules.mla. They are not re-exported here to avoid circular imports:
+# custom_ops must not depend on modules.attention or modules.mla.
 
 
 def inplace_slice_copy(dest: torch.Tensor, src: torch.Tensor, dim1_start: int,
