+++ tensorrt_llm/_torch/modules/ATTENTION_DEVELOPER_GUIDE.md
@@ -5,6 +5,7 @@
 This guide covers the TRT-LLM PyTorch attention stack:
 
 - `tensorrt_llm/_torch/modules/attention.py`
+- `tensorrt_llm/_torch/modules/mla.py`
 - `tensorrt_llm/_torch/attention_backend/`
 - `tensorrt_llm/_torch/attention_backend/sparse/`
 
@@ -92,7 +93,7 @@ without changing the outer runtime contract.
 
 ### 1.2 `MLA`: a separate module on top of the same backend system
 
-`MLA` (Multi-head Latent Attention) is a separate module in `attention.py`.
+`MLA` (Multi-head Latent Attention) is a separate module in `mla.py`.
 Like `Attention`, it keeps module-level projection logic in the module,
 delegates core execution to a backend object, and depends on metadata and
 KV-cache contract. At a high level, it owns:
@@ -116,7 +117,7 @@ For MLA-related tasks, first check whether the work fits the current
 projection structure, can stay on an existing backend and metadata family, and
 can preserve the current latent-cache / paged-KV contract. If it can, the
 task usually stays within the existing MLA stack. If it depends on sparse
-helper-level control flow, read `attention.py` and the relevant sparse
+helper-level control flow, read `mla.py` and the relevant sparse
 backend code directly.
 
 ## 2. Backend Layer Reference
@@ -278,7 +279,7 @@ appending, RoPE application, and loading cached state for attention use.
 
 MLA fit cannot be judged from attention math alone. The module and backend must
 agree on latent-cache layout, paged-KV read/write paths, and cached/chunked
-context behavior. Read the MLA section of `attention.py` and the relevant
+context behavior. Read `mla.py` and the relevant
 backend code for the current implementation details.
 
 #### 3.2.4 Sparse side-cache semantics
@@ -353,7 +354,8 @@ Working rules:
 
 | File | Role |
 |------|------|
-| `tensorrt_llm/_torch/modules/attention.py` | Standard attention and MLA module logic |
+| `tensorrt_llm/_torch/modules/attention.py` | Standard attention module logic and shared Helix CP helpers |
+| `tensorrt_llm/_torch/modules/mla.py` | MLA module logic, MLA custom ops, and MLA-specific dispatch |
 | `tensorrt_llm/_torch/attention_backend/interface.py` | Backend contract, base metadata, capability hooks |
 | `tensorrt_llm/_torch/attention_backend/utils.py` | Backend and sparse-backend selection |
 | `tensorrt_llm/_torch/attention_backend/trtllm.py` | TRTLLM backend and metadata |
