+++ tensorrt_llm/tools/layer_wise_benchmarks/mark_utils.py
@@ -6,10 +6,11 @@ from tensorrt_llm._torch.models.modeling_qwen3_next import (
     Qwen3NextGatedDeltaNet,
     Qwen3NextSparseMoeBlock,
 )
-from tensorrt_llm._torch.modules.attention import MLA, Attention
+from tensorrt_llm._torch.modules.attention import Attention
 from tensorrt_llm._torch.modules.fused_moe.interface import MoE
 from tensorrt_llm._torch.modules.gated_mlp import GatedMLP
 from tensorrt_llm._torch.modules.mamba.mamba2_mixer import Mamba2Mixer
+from tensorrt_llm._torch.modules.mla import MLA
 
 
 def mark_ranges():
